View-dep cleanups (really paper overs, code needs a rewrite)
Tighter check for ref type in resolver (this was broken but unnoticed)
Skip table result classes in view part of resolver.
No uninitialized warnings in resolver where it expects some misses (ugh!)
Cleanup viewdeps test (and correct incorrect comparison)

This should have not merged as is, but time is short to fix it properly.
Hopefully next iter.