Make Data::Page total count evaluation lazy
The short version is - now you can do $rs->pager->total_entries($satanic_num)
and DBIC will never even try to fire a COUNT query

This may seem to be a *lot* of code to do something that could
be done with a trivial subclass. However keep in mind that this
is a proof-of-concept of what do we do when we can *not* (or do
not want to) subclass.