Fix $object->search_related aliasing, change semantics of _resolve_condition
Change the RV of _resolve_condition one last time. Now it checks the RV of the
resolved coderef (if any), and chooses (preferrably) the join-free condition
or the cross-table one. In list context returns a flag signifying if the
resulting condition is available for standalone use (false) or requires the
joins to remain (true)