prefetch docs: prefetch works on all rel types
Update the prefetch docs to state that prefetch works with all
relationship types, where previously it had said that it works with
only belongs_to and has_one.

Also update the subsequent example. The old example was confusing
because it specified two has_many relationships on the same level,
something that DBIC emits a warning about.  That example is replaced
with a CD-centric example that demonstrates how many different
relationship types can be used together in a prefetch, but omits the
multiple has_manys.

Add a warning, example, and explanation of the troubles with using
multiple has_many relations in a prefetch.

The old text incorrectly stated that prefetch will override join and
select attributes. Describe current behavior with examples.