fix a couple test failures on Win32
Adds a skip_all to t/storage/global_destruction.t due to it causing a
segfault triggered by some sort of perl bug.

Adds a \Q...\E to t/94versioning.t in a warning regex as well as
changing a `perl -pi -e` to `perl -pi.bak -e` and removing an extraneous
semicolon for the test to pass.