fix popups during tests on Win32 (hopefully)
A tester reports that t/73oracle_hq.t causes an error popup on Win32.
This test was checking for DBD::Oracle (and thereby trying to load it)
before checking for the DBICTEST_ORA_DSN environment variables.

This commit moves the check for necessary environment variables for the
t/73oracle.t and t/73oracle_hq.t test files to before the module
requirements tests, thereby not trying to load DBD::Oracle, and causing
an error popup.