Fix TxnScopeGuard misbehaving on externally set $@ without inner exceptions
The guard is supposed to warn when it goes out of scope without a commit.
However this doesn't work if $@ was already set before the guard was created
and there were no eval{}s to clear it between definition and destruction.

Fixing this by storing a weakref to the current exception (if any) and
disregarding the contents of $@ if they match the ref we started with.