fix Win32 file use error in t/storage/reconnect.t
Even though the fake SQLite db file is chmod 0000, for some reason
SQLite still opens it on Win32 and holds it open; this prevents deleting
it and replacing it with the original db file so that the rest of the
tests can run.

TODO: followup with the DBD::SQLite folks.

Disconnect the $dbh on Win32 so the unlink can succeed, and remove the
following SKIP: block (for if db is "encrypted") which is no longer
necessary.