add ::DBI::_is_integer_type and use for SQLite
Add _is_integer_type to check whether a data_type is an integer type
from any database, like we already have _is_lob_type. Use it for the
SQL_INTEGER bind in the SQLite driver.

Also for ::SQLite check for a generic numeric type with a zero scale
such as NUMBER(38,0) and make it a SQL_INTEGER bind.

Move the ::SQLite integer type checks into _resolve_bindattrs in order
to cache the result of the check in the column_info, since the regexes
are very hairy.