skip 51threadnodb.t on Win32 on < 5.14
SineSwiper reported a crash with a dialog from "make test" in
t/51threadnodb.t on Strawberry Perl 5.12 32bit.

The test passes on my ActivePerl 64bit 5.14.2 and there is a passing
test report for 5.14 here:
http://www.cpantesters.org/cpan/report/239d2ae4-87c3-1015-ad08-a09405085451

Disable test on Win32 for perls below 5.14 unless the environment
variable TEST_AUTHOR is set to be safe.