Rewrite 461c7168 solving Oracle silent connection failures
The patch incorrectly masks the underlying issue - during the refactor
of 584ea6e4 a bunch of exception traps were introduced which were never
there in the first place. _dbh_get_info() should always be throwing an
error if one takes place (the only place before 584ea6e4 was in the
FreeTDS version determination, and that one made no sense anyway).
_server_info() OTOH is indeed an opportunistic routine - leave the
"rethrow if determining driver" escape as-is