Fix very subtle but deadly bug in single()
The _select_args result storage introduced in 975b573a did not take
into account that single() operates on the $rs itself, not a fresh
one. As such the first single() arg would be carried around and then
applied to subsequent next() etc iterations. Stop that, as we only
need the attrs in case of collapse => 1 and single() explicitly does
not support this.

While at it do not reuse the cached args at all - while it *appears*
safe, it may very well not be. Leave this for another day