Examples and documentation update and cleanup
This version updates the original POD with slightly more information
so that more DBIx::Class functionality is revealed.  The source code is
removed from POD, and instead the reader is directed to the examples
directory.

It also changes the example schema from silly auto-inferred column
names to proper *id-suffixed ones.