Documentation update and cleanup
This update removes 'year' from the CD table definition in both the SQL
and in the class definition, since it wasn't being used. The directory
MyApp/Schema/ResultSet was also removed since the supplied code never
used it.

The results that the sample code produces have been expanded a little so
that they provide answers to the questions asked by the subroutine
calls.

A new section has been added that briefly discusses the results and
explains the techniques that could be used when writing this type of a
programmatic query.