Fix exceedingly sloppy SQLite warning workaround from 04ab4eb1
The previous version managed not only to *not* silence any warning but also
break the test checking for lack of said warnings. In addition the silencing
regex was unacceptably broad, possibly missing some important misbinds.

riba pull your shit together man
a.k.a. this is what happens when there are no code reviewers :(((